test for new feature

+++ server.py
@@ -179,7 +179,10 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'username' not in login_session:
+    return render_template('publicRestaurants.html', restaurants = restaurants)
+  else:
+    return render_template('restaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -201,6 +204,8 @@ def editRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if editedRestaurant.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()'>"
   if request.method == 'POST':
       if request.form['name']:
         editedRestaurant.name = request.form['name']
@@ -216,6 +221,8 @@ def deleteRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if restaurantToDelete.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()'>"
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -229,8 +236,12 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+      return render_template('publicMenu.html', items=items, restaurant=restaurant, creator=creator)
+    else:
+      return render_template('menu.html', items = items, restaurant = restaurant, creator = creator)
      
 
 
@@ -240,6 +251,8 @@ def newMenuItem(restaurant_id):
   if 'username' not in login_session:
       return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if login_session['user_id'] != restaurant.user_id:
+    return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()''>"
   if request.method == 'POST':
       newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id, user_id = restaurant.user_id)
       session.add(newItem)
@@ -256,19 +269,21 @@ def editMenuItem(restaurant_id, menu_id):
       return redirect('/login')
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+      return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
-        if request.form['name']:
+      if request.form['name']:
             editedItem.name = request.form['name']
-        if request.form['description']:
+      if request.form['description']:
             editedItem.description = request.form['description']
-        if request.form['price']:
+      if request.form['price']:
             editedItem.price = request.form['price']
-        if request.form['course']:
+      if request.form['course']:
             editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+      session.add(editedItem)
+      session.commit() 
+      flash('Menu Item Successfully Edited')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
         return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
@@ -280,13 +295,15 @@ def deleteMenuItem(restaurant_id,menu_id):
       return redirect('/login')
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
+    if login_session['user_id'] != restaurant.user_id:
+      return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+      session.delete(itemToDelete)
+      session.commit()
+      flash('Menu Item Successfully Deleted')
+      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
     else:
-        return render_template('deleteMenuItem.html', item = itemToDelete, restaurant_id = restaurant_id)
+      return render_template('deleteMenuItem.html', item = itemToDelete, restaurant_id = restaurant_id)
 
 # User Helper Functions
 
@@ -298,7 +315,7 @@ def getUserID(email):
     return None
 
 def getUserInfo(user_id):
-  user = session.query(User).fliter_by(id=user_id).one()
+  user = session.query(User).filter_by(id=user_id).one()
   return user
 
 def createUser(login_session):
